author+access+Personalization=Merger+analysis
Once the owner is ready and his project to have changes from the food to the non-limited change without breaking the taken repository has a new function without requesting the technical and other departments. I'll call it the AS/project button asking you to launch this improvement on my behalf as an author. The button accompanies the action of any new resource and is equipped with security and requirements passing which is analyzed with a period of twenty four hours Combining the common effort by building a common repository by removing repetitions of similarities and errors after which twenty-four hours later there are reports to participants personally with their contribution and further action even if not willing to work in the project the person receives an award as a diploma and a description of merit for motivating to take into account the units of each award without repeating ever and Having an exchange back as a sale but the possibility of further participation. Thanks hope to appreciate the above written implementation as an idea.